requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
django-filter==23.4
Pillow==10.0.1
python-decouple==3.8
psycopg2-binary==2.9.9
PyJWT==2.10.1
django-jwt-extended==2.0.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
celery==5.3.4
redis==5.0.1
dj-database-url==2.1.0
